backend, front end: improve token handling
Add user id in token and add get session endpoint to be able to get
current user from a cookie (containing the token), and for the front to
be able to determine if a valid sesson is available.

Return the user in the login endpoint too.

related feature: user login